login system implemented, and minor css changes

+++ sifsearchdemo/sifsearchdemo/views.py
@@ -1,5 +1,5 @@
 from django.http import JsonResponse
-from django.shortcuts import render
+from django.shortcuts import render,redirect
 from .models import SearchEntry, Tag
 from .serializers import EntrySerializer,EntryFileSerializer, TagSerializer
 from rest_framework.decorators import api_view, permission_classes
@@ -9,8 +9,17 @@ from rest_framework import status
 from django.views.decorators.csrf import csrf_exempt
 from .models import Tag
 import algoliasearch_django as algoliasearch
+from .forms import NewUserForm
+from django.contrib.auth import login
+from django.contrib import messages
+from django.contrib.auth import login, authenticate
+from django.contrib.auth.forms import AuthenticationForm
+from django.contrib.auth import logout
+
 
 def sif_search_home(request):
+    if not request.user.is_authenticated:
+        return redirect("/login")
     currtags = Tag.objects.all().values()
     algoliasearch.reindex_all(SearchEntry,) 
     return render(request, 'home.html', {'currtags':currtags})
@@ -21,6 +30,14 @@ def add_entry_link(request):
     serializer = EntrySerializer(data=request.data)
     tag_serializer = TagSerializer(data=request.data)
     tag_name = request.POST.get('tag')
+
+
+    '''
+    basic request (raw)--> {
+        name: value,
+        tag: value
+    }
+    '''
 
     if serializer.is_valid():
         # testing to see if this will work
@@ -61,19 +78,61 @@ def add_entry_file(request):
 @api_view(['PUT'])
 def update_entry(request):
     entry_name = request.POST.get('name')
+    user_name = request.POST.get('user')
     entry_obj = SearchEntry.objects.get(name=entry_name) # search for the original entry using the name
+    print(entry_obj.user, user_name)
+    if entry_obj.user == user_name:
     # modify the name and description
-    entry_obj.name = request.POST.get('new-name')
-    entry_obj.description = request.POST.get('new-description')
-    entry_obj.save() # saving the changes
-
-    return Response(status=status.HTTP_201_CREATED)
+        entry_obj.name = request.POST.get('new-name')
+        entry_obj.description = request.POST.get('new-description')
+        entry_obj.save() # saving the changes
+        return Response(status=status.HTTP_201_CREATED)
+    return Response(status=status.HTTP_406_NOT_ACCEPTABLE)
 
 
 # function for deleting a SearchEntry
 @api_view(['DELETE'])
 def delete_entry(request):
     entry_name = request.POST.get('name')
-    SearchEntry.objects.filter(name = entry_name).delete()
-    return Response(status=status.HTTP_201_CREATED)
+    user_name = request.POST.get('user')
+    if SearchEntry.objects.get(name=entry_name).user == user_name:
+        SearchEntry.objects.filter(name = entry_name).delete()
+        return Response(status=status.HTTP_201_CREATED)
+    return Response(status=status.HTTP_406_NOT_ACCEPTABLE)
 
+
+
+# function for registering a new user into the site
+def register_request(request):
+	if request.method == "POST":
+		form = NewUserForm(request.POST)
+		if form.is_valid():
+			user = form.save()
+			login(request, user)
+			messages.success(request, "Registration successful." )
+			return redirect("/login")
+		messages.error(request, "Unsuccessful registration. Invalid information.")
+	form = NewUserForm()
+	return render (request=request, template_name="register.html", context={"register_form":form})
+
+def login_request(request):
+	if request.method == "POST":
+		form = AuthenticationForm(request, data=request.POST)
+		if form.is_valid():
+			username = form.cleaned_data.get('username')
+			password = form.cleaned_data.get('password')
+			user = authenticate(username=username, password=password)
+			if user is not None:
+				login(request, user)
+				messages.info(request, "You are now logged in as {username}.")
+				return redirect("/")
+			else:
+				messages.error(request,"Invalid username or password.")
+		else:
+			messages.error(request,"Invalid username or password.")
+	form = AuthenticationForm()
+	return render(request=request, template_name="login.html", context={"login_form":form})
+def logout_request(request):
+    logout(request)
+    messages.info(request, "Logged out successfully!")
+    return redirect("login")
